Create IndividualDetails component
