+++ frontend/src/components/pages/IntakePage.tsx
@@ -80,12 +80,7 @@ const Intake = (): React.ReactElement => {
         return (
           <>
             <Box style={{ textAlign: "center", padding: "30px 0px 40px 0px" }}>
-              <ReviewForm
-                prevStep={prevStep}
-                title="Individual details"
-                childrenDetails={[]}
-                caregiverDetails={[]}
-              />
+              <ReviewForm prevStep={prevStep} />
             </Box>
           </>
         );
